Add AWS S3 package and configure file storage
This commit adds the AWSSDK.S3 package to the project and configures file storage using Amazon S3. The `AWSSDK.S3` package is added as a reference in the `Directory.Packages.props` and `Infrastructure.csproj` files. The `DependencyInjection.cs` file is modified to include the necessary services for file storage using Amazon S3. The `AddFileStorage` extension method is added to the `DependencyInjection` class to configure the AWS S3 settings and register the `IAmazonS3` client. The `AwsS3Settings` class is added to the `Application/Configurations` folder to hold the AWS S3 settings. The `IStorageManagementService` interface is added to the `Application/Interfaces` folder, and the `AwsS3Service` class is added to the `Infrastructure/Service` folder to implement the file storage functionality using Amazon S3. The `GenerateUrlToUploadFile` command, validator, and handler classes are added to the `Application/FileManagements/Commands/GenerateUrlToUploadFile` folder to handle generating a URL for file upload. The `StorageManagements` endpoint group is added to the `Web/Endpoints` folder to handle the file management API endpoints.